# back/modules/generated/public/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "Chaman",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://chaman.frol.be/api"
  }, {
    "url" : "http://chaman.frol.be/api"
  } ],
...
  "paths" : {
...
    "/item" : {
      "get" : {
        "operationId" : "getItems",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ItemDescr"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get all items",
        "tags" : [ "item" ]
      },
      "post" : {
        "operationId" : "createItem",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Item"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Item"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "create an item",
        "tags" : [ "item" ]
      }
    },
    "/item/{uuid}" : {
      "delete" : {
        "operationId" : "deleteItem",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "uuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : { },
            "description" : "OK"
          }
        },
        "summary" : "delete a item",
        "tags" : [ "item" ]
      },
      "get" : {
        "operationId" : "getItem",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "uuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Item"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "get a item",
        "tags" : [ "item" ]
      },
      "put" : {
        "operationId" : "updateItem",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "uuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Item"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Item"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "update a item",
        "tags" : [ "item" ]
      }
    },
...
  },
  "components" : {
...
    "schemas" : {
...
      "Field" : {
        "example" : {
          "reference" : "reference",
          "errorMessages" : [ "errorMessages", "errorMessages" ],
          "inputType" : "inputType",
          "label" : "label",
          "params" : "",
          "uuid" : "uuid",
          "value" : [ {
            "uuid" : "uuid",
            "value" : ""
          }, {
            "uuid" : "uuid",
            "value" : ""
          } ]
        },
        "properties" : {
          "uuid" : {
            "description" : "single uuid acrross all version of the field",
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "reference" : {
            "type" : "string"
          },
          "inputType" : {
            "type" : "string"
          },
          "value" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldValue"
            },
            "type" : "array"
          },
          "params" : { },
          "errorMessages" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "FieldValue" : {
        "example" : {
          "uuid" : "uuid",
          "value" : ""
        },
        "properties" : {
          "uuid" : {
            "description" : "identifier of the value",
            "type" : "string"
          },
          "value" : { }
        },
        "type" : "object"
      },
      "ItemDescr" : {
        "example" : {
          "description" : "description",
          "title" : "title",
          "uuid" : "uuid"
        },
        "properties" : {
          "uuid" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Item" : {
        "example" : {
          "description" : "description",
          "links" : [ {
            "item" : {
              "description" : "description",
              "title" : "title",
              "uuid" : "uuid"
            },
            "uuid" : "uuid"
          }, {
            "item" : {
              "description" : "description",
              "title" : "title",
              "uuid" : "uuid"
            },
            "uuid" : "uuid"
          } ],
          "title" : "title",
          "uuid" : "uuid",
          "annexes" : [ {
            "filename" : "filename",
            "mime" : "mime",
            "uuid" : "uuid"
          }, {
            "filename" : "filename",
            "mime" : "mime",
            "uuid" : "uuid"
          } ],
          "content" : [ {
            "reference" : "reference",
            "errorMessages" : [ "errorMessages", "errorMessages" ],
            "inputType" : "inputType",
            "label" : "label",
            "params" : "",
            "uuid" : "uuid",
            "value" : [ {
              "uuid" : "uuid",
              "value" : ""
            }, {
              "uuid" : "uuid",
              "value" : ""
            } ]
          }, {
            "reference" : "reference",
            "errorMessages" : [ "errorMessages", "errorMessages" ],
            "inputType" : "inputType",
            "label" : "label",
            "params" : "",
            "uuid" : "uuid",
            "value" : [ {
              "uuid" : "uuid",
              "value" : ""
            }, {
              "uuid" : "uuid",
              "value" : ""
            } ]
          } ]
        },
        "properties" : {
          "uuid" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/Field"
            },
            "type" : "array"
          },
          "annexes" : {
            "items" : {
              "$ref" : "#/components/schemas/Annex"
            },
            "type" : "array"
          },
          "links" : {
            "items" : {
              "$ref" : "#/components/schemas/Link"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "Annex" : {
        "example" : {
          "filename" : "filename",
          "mime" : "mime",
          "uuid" : "uuid"
        },
        "properties" : {
          "uuid" : {
            "type" : "string"
          },
          "filename" : {
            "type" : "string"
          },
          "mime" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Link" : {
        "example" : {
          "item" : {
            "description" : "description",
            "title" : "title",
            "uuid" : "uuid"
          },
          "uuid" : "uuid"
        },
        "properties" : {
          "uuid" : {
            "type" : "string"
          },
          "item" : {
            "$ref" : "#/components/schemas/ItemDescr"
          }
        },
        "type" : "object"
      },
...
    }
  }
}